perf: Speed up sync index database
Fix --wait-for-sync
