+++ src/utils/mod.rs
@@ -2,6 +2,7 @@ pub mod arg;
 pub mod arg_parser;
 pub mod completer;
 pub mod config;
+pub mod index;
 pub mod json_color;
 pub mod other;
 pub mod printer;
